fix. deleted comments
