+++ src/app/features/public/pages/activities/components/activities-card/activities-card.component.html
